printing messages from chat
