+++ cmd/server/main.go
@@ -1,16 +1,16 @@
 package main
 
 import (
-	"fmt"
+
 	// "log"
-	"net/http"
+
 	// "strings"
+	"fmt"
 	"time"
 
-	// client "github.com/fomiller/poggers/pkg/twitch"
 	// twitch "github.com/gempir/go-twitch-irc/v3"
 
-	"github.com/gin-gonic/gin"
+	"github.com/gempir/go-twitch-irc/v3"
 	"github.com/gorilla/websocket"
 )
 
@@ -34,42 +34,73 @@ var (
 )
 
 func main() {
-	r := gin.Default()
-	r.LoadHTMLFiles("./cmd/server/index.html")
-	r.GET("/", homeHandler)
-	r.GET("/ws", websocketHandler)
+	// r := gin.Default()
+	// r.LoadHTMLFiles("./cmd/server/index.html")
+	// r.GET("/", homeHandler)
+	// r.GET("/ws", websocketHandler)
 
-	r.Run(":8080")
-}
+	// r.Run(":8080")
 
-func homeHandler(c *gin.Context) {
-	c.HTML(200, "index.html", nil)
-}
+	client := twitch.NewAnonymousClient()
 
-func websocketHandler(c *gin.Context) {
-	wsHandler(c.Writer, c.Request)
-}
+	client.OnPrivateMessage(func(message twitch.PrivateMessage) {
+		// if strings.Contains(strings.ToLower(message.Message), "pog") {
+		// 	x := PogMessage{message.User.Name, message.Message, message.Time}
+		// 	// pogList = append(pogList, x)
 
-func wsHandler(w http.ResponseWriter, r *http.Request) {
-	conn, err := upgrader.Upgrade(w, r, nil)
-	if err != nil {
-		fmt.Printf("Failed to set websocket upgrade: %+v", err)
-		return
-	}
-	fmt.Println(r.URL.RequestURI())
-	for {
-		// if err != nil {
-		// 	break
+		// 	// pogCount++
+		// 	fmt.Println(x)
+
+		// 	// client.Say(Channel, fmt.Sprintf("Pog has been said %v times", pogCount))
 		// }
-		// print(t)
-		msg := "hello"
-		t := 1
-		conn.WriteMessage(t, []byte(fmt.Sprintf("SUP %s", msg)))
-		time.Sleep(time.Second)
+		fmt.Println(message.Message)
+	})
+
+	client.Join("swolenesss")
+
+	err := client.Connect()
+	if err != nil {
+		panic(err)
 	}
 }
 
-// func ReadChat(client *twitch.Client) {
+// func homeHandler(c *gin.Context) {
+// 	c.HTML(200, "index.html", nil)
+// }
+
+// func websocketHandler(c *gin.Context) {
+// 	client := twitch.NewAnonymousClient()
+
+// 	client.Join("swolenesss")
+
+// 	err := client.Connect()
+// 	if err != nil {
+// 		panic(err)
+// 	}
+// 	wsHandler(c.Writer, c.Request, client)
+// }
+
+// func wsHandler(w http.ResponseWriter, r *http.Request, c *twitch.Client) {
+// 	conn, err := upgrader.Upgrade(w, r, nil)
+// 	if err != nil {
+// 		fmt.Printf("Failed to set websocket upgrade: %+v", err)
+// 		return
+// 	}
+// 	fmt.Println(r.URL.RequestURI())
+// 	go ReadChat(c, conn)
+// for {
+// 	// if err != nil {
+// 	// 	break
+// 	// }
+// 	// print(t)
+// 	msg := "hello"
+// 	t := 1
+// 	conn.WriteMessage(t, []byte(fmt.Sprintf("sup %s", msg)))
+// 	time.Sleep(time.Second)
+// }
+// }
+
+// func ReadChat(client *twitch.Client, conn *websocket.Conn) {
 // 	client.OnPrivateMessage(func(message twitch.PrivateMessage) {
 // 		if strings.Contains(strings.ToLower(message.Message), "pog") {
 // 			x := PogMessage{message.User.Name, message.Message, message.Time}
@@ -79,7 +110,8 @@ func wsHandler(w http.ResponseWriter, r *http.Request) {
 // 			fmt.Println(pogCount)
 
 // 			client.Say(Channel, fmt.Sprintf("Pog has been said %v times", pogCount))
+// 			t := 1
+// 			conn.WriteMessage(t, []byte(fmt.Sprintf("sup %s", message.Message)))
 // 		}
 // 	})
-
 // }
